requirements.txt
bottle==0.12.19
colorama==0.4.4
python-dotenv==0.19.2
waitress==2.0.0
